refactor: base.ts -> llm.ts
